+++ lib/generators/show_for/install_generator.rb
@@ -2,7 +2,7 @@ module ShowFor
   module Generators
     class InstallGenerator < Rails::Generators::Base
       desc "Copy ShowFor installation files"
-      class_option :template_engine, :desc => 'Template engine to be invoked (erb or haml or slim).'
+      class_option :template_engine, desc: 'Template engine to be invoked (erb or haml or slim).'
       source_root File.expand_path('../templates', __FILE__)
 
       def copy_initializers