+++ test/attribute_test.rb
@@ -5,7 +5,7 @@ class AttributeTest < ActionView::TestCase
   test "show_for accepts an attribute as a collection" do
     with_attribute_for @user, :scopes
     assert_select "div.show_for div.wrapper ul.collection"
-    assert_select "div.show_for div.wrapper ul.collection li", :count => 3
+    assert_select "div.show_for div.wrapper ul.collection li", count: 3
   end
 
   test "show_for accepts an attribute as a collection with a block to iterate the collection" do
@@ -13,25 +13,25 @@ class AttributeTest < ActionView::TestCase
       content_tag :span, scope
     end
     assert_select "div.show_for div.wrapper ul.collection"
-    assert_select "div.show_for div.wrapper ul.collection span", :count => 3
+    assert_select "div.show_for div.wrapper ul.collection span", count: 3
   end
 
   test "show_for treats symbol for :value as method on each element of collection" do
-    with_attribute_for @user, :scopes, :value => :upcase
+    with_attribute_for @user, :scopes, value: :upcase
     @user.scopes.each do |scope|
       assert_select "div.show_for div.wrapper ul.collection", /#{scope.upcase}/
     end
   end
 
   test "show_for allows collection tag to be configured globally" do
-    swap ShowFor, :collection_tag => :ol, :collection_class => "my_collection" do
+    swap ShowFor, collection_tag: :ol, collection_class: "my_collection" do
       with_attribute_for @user, :scopes
       assert_select "div.show_for div.wrapper ol.my_collection"
     end
   end
 
   test "show_for allows collection class to be disabled globally" do
-    swap ShowFor, :collection_tag => :ol, :collection_class => nil do
+    swap ShowFor, collection_tag: :ol, collection_class: nil do
       with_attribute_for @user, :scopes
       assert_select "div.show_for div.wrapper ol"
       assert_no_select "ol[class]"
@@ -39,25 +39,25 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for allows collection tag to be changed by attribute" do
-    with_attribute_for @user, :scopes, :collection_tag => :ol
+    with_attribute_for @user, :scopes, collection_tag: :ol
     assert_select "div.show_for div.wrapper ol.collection"
   end
 
   test "show_for allows collection tag html to be configured by attribute" do
-    with_attribute_for @user, :scopes, :collection_html => { :id => "thecollection", :class => "special" }
+    with_attribute_for @user, :scopes, collection_html: { id: "thecollection", class: "special" }
     assert_select "div.show_for div.wrapper ul#thecollection.special.collection"
   end
 
   # CONTENT
   test "show_for allows content tag to be configured globally" do
-    swap ShowFor, :content_tag => :span, :content_class => :my_content do
+    swap ShowFor, content_tag: :span, content_class: :my_content do
       with_attribute_for @user, :name
       assert_select "div.show_for div.wrapper span.my_content"
     end
   end
 
   test "show_for allows content class to be disabled globally" do
-    swap ShowFor, :content_tag => :span, :content_class => nil do
+    swap ShowFor, content_tag: :span, content_class: nil do
       with_attribute_for @user, :name
       assert_select "div.show_for div.wrapper span"
       assert_no_select "span[class]"
@@ -65,12 +65,12 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for allows content tag to be changed by attribute" do
-    with_attribute_for @user, :name, :content_tag => :span
+    with_attribute_for @user, :name, content_tag: :span
     assert_select "div.show_for div.wrapper span.content"
   end
 
   test "show_for allows content tag html to be configured by attribute" do
-    with_attribute_for @user, :name, :content_tag => :span, :content_html => { :id => "thecontent", :class => "special" }
+    with_attribute_for @user, :name, content_tag: :span, content_html: { id: "thecontent", class: "special" }
     assert_select "div.show_for div.wrapper span#thecontent.special.content"
   end
 
@@ -90,8 +90,8 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for accepts an attribute as time with format options" do
-    with_attribute_for @user, :created_at, :format => :long
-    assert_select "div.show_for div.wrapper", /#{Regexp.escape(I18n.l(@user.created_at, :format => :long))}/
+    with_attribute_for @user, :created_at, format: :long
+    assert_select "div.show_for div.wrapper", /#{Regexp.escape(I18n.l(@user.created_at, format: :long))}/
   end
 
   test "show_for accepts an attribute as true" do
@@ -100,7 +100,7 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for accepts an attribute as true which can be localized" do
-    store_translations(:en, :show_for => { :yes => "Hell yeah!" }) do
+    store_translations(:en, show_for: { yes: "Hell yeah!" }) do
       with_attribute_for @user, :active
       assert_select "div.show_for div.wrapper", /Hell yeah!/
     end
@@ -112,7 +112,7 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for accepts an attribute as false which can be localized" do
-    store_translations(:en, :show_for => { :no => "Hell no!" }) do
+    store_translations(:en, show_for: { no: "Hell no!" }) do
       with_attribute_for @user, :invalid
       assert_select "div.show_for div.wrapper", /Hell no!/
     end
@@ -124,21 +124,21 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for accepts not spcified message can be localized" do
-    store_translations(:en, :show_for => { :blank => "OMG! It's blank!" }) do
+    store_translations(:en, show_for: { blank: "OMG! It's blank!" }) do
       with_attribute_for @user, :description
       assert_select "div.show_for div.wrapper", /OMG! It's blank!/
     end
   end
 
   test "show_for accepts not spcified message can be localized with html" do
-    store_translations(:en, :show_for => { :blank_html => "<span>OMG! It's blank!</span>" }) do
+    store_translations(:en, show_for: { blank_html: "<span>OMG! It's blank!</span>" }) do
       with_attribute_for @user, :description
       assert_select "div.show_for div.wrapper span", "OMG! It's blank!"
     end
   end
 
   test "show_for uses :if_blank if attribute is blank" do
-    with_attribute_for @user, :description, :if_blank => "No description provided"
+    with_attribute_for @user, :description, if_blank: "No description provided"
     assert_select "div.show_for div.wrapper", /No description provided/
   end
 
@@ -150,7 +150,7 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for uses :if_blank if the block content is blank" do
-    with_attribute_for @user, :description, :if_blank => "No description provided" do
+    with_attribute_for @user, :description, if_blank: "No description provided" do
       ""
     end
     assert_select "div.show_for div.wrapper", /No description provided/
@@ -178,41 +178,41 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for uses :value if supplied" do
-    with_attribute_for @user, :name, :value => "Calculated Value"
+    with_attribute_for @user, :name, value: "Calculated Value"
     assert_select "div.show_for div.wrapper", /Calculated Value/
   end
 
   test "show_for uses :value and casts to string if supplied" do
-    with_attribute_for @user, :name, :value => 123
+    with_attribute_for @user, :name, value: 123
     assert_select "div.show_for div.wrapper", /123/
   end
 
   test "show_for ignores attribute if :value supplied but with nil value" do
-    with_attribute_for @user, :name, :value => nil
+    with_attribute_for @user, :name, value: nil
     assert_select "div.show_for div.wrapper", /Not specified/
   end
 
   test "show_for ignores :value if a block is supplied" do
-    with_attribute_for @user, :name, :value => "Calculated Value" do
+    with_attribute_for @user, :name, value: "Calculated Value" do
       @user.name.upcase
     end
     assert_select "div.show_for div.wrapper", /#{@user.name.upcase}/
   end
 
   test "show_for treats symbol for :value as method on attribute" do
-    with_attribute_for @user, :name, :value => :upcase
+    with_attribute_for @user, :name, value: :upcase
     assert_select "div.show_for div.wrapper", /#{@user.name.upcase}/
   end
 
   test "show_for does not display blank attribute if skip_blanks option is passed" do
-    swap ShowFor, :skip_blanks => true do
+    swap ShowFor, skip_blanks: true do
       with_attribute_for @user, :description
       assert_no_select "div.show_for div.wrapper"
     end
   end
 
   test "show_for display false attribute if skip_blanks option is passed" do
-    swap ShowFor, :skip_blanks => true do
+    swap ShowFor, skip_blanks: true do
       with_attribute_for @user, :invalid
       assert_select "div.show_for div.wrapper", /No/
     end
@@ -228,7 +228,7 @@ class AttributeTest < ActionView::TestCase
   end
 
   test "show_for should wrap blank attributes with no_attribute" do
-    swap ShowFor, :blank_content_class => 'no_attribute' do
+    swap ShowFor, blank_content_class: 'no_attribute' do
       with_attributes_for @user, :name, :birthday, :karma
       assert_select ".wrapper.user_birthday.no_attribute"
       assert_select ".wrapper.user_karma.no_attribute"